Added missing hud stuff to 2003. hope it all works!
This is so dexter (& shok?) can help with the following:

buildstate_sg is a buildstate is a quantitypanel
quantitypanel has children as quantitybar
buildstate has cvars to define the style of the children

As it stands now buildstate checks the cvars every other tick to see if they've changed from the ones saved as members
if they've changed then the member gets updated and it updates the child quantity bars
theres a lot of the same variables

I've had a thought, not sure if it's worth doing depending on how I do this whole thing but would solve the duplicated member problem: Have a style class that the quantity bars are given a pointer to, the buildstate class updates the style object based on the cvars.

I just need a better way to update the styles.

Side note: The color fading is broken in some way at the moment but I'm not sure how. It might just be because the SG goes above 100 and the color scale hasn't adapted for that - it's something I am aware of and need to fix.

# Conflicts:
#	cl_dll/client_scratch-2003.vcproj
#	cl_dll/client_scratch-2005.vcproj
#	cl_dll/client_scratch-2008.vcproj