Changing tick signals. Previously every tick to every 100ms (good news for 33 tick and more so 100 tick servers!)
hud_crosshair to 500 as its only used right at the start of the game to load the players crosshair preference and then removed. Not important at all but meh, no point in it doing it too often.

StatusIcons and rootpanel have had tick signals removed or commented as they are not used.

# Conflicts:
#	cl_dll/ff/ff_hud_crosshairinfo.cpp