Custom hud update
No point putting all changes in as there's so many.
Added preview checkbox - might not be clear how this works and end up having preview enabled all the time.
Added anchor positions for text to allow for more flexibility.
Need to add control tooltips in a future update to make the controls that bit clearer.

Still more changes planned but settled on and fixed much of the base functionality!

# Conflicts:
#	Game_Scratch-2008.sln
#	cl_dll/client_scratch-2008.vcproj
#	cl_dll/ff/vgui/ff_miscoptions.h
#	cl_dll/vgui_int.cpp
#	dlls/server_scratch-2008.vcproj